+++ packages/components/radio/__tests__/radio.test.tsx
@@ -144,9 +144,9 @@ describe("Radio", () => {
     expect(onFocus).toBeCalled();
   });
 
-  it('should work correctly with "isRequired" prop', () => {
+  it("should have required attribute when isRequired with native validationBehavior", () => {
     const {getByRole, getAllByRole} = render(
-      <RadioGroup isRequired label="Options">
+      <RadioGroup isRequired label="Options" validationBehavior="native">
         <Radio value="1">Option 1</Radio>
         <Radio value="2">Option 2</Radio>
       </RadioGroup>,
@@ -159,6 +159,23 @@ describe("Radio", () => {
     const radios = getAllByRole("radio");
 
     expect(radios[0]).toHaveAttribute("required");
+  });
+
+  it("should not have required attribute when isRequired with aria validationBehavior", () => {
+    const {getByRole, getAllByRole} = render(
+      <RadioGroup isRequired label="Options" validationBehavior="aria">
+        <Radio value="1">Option 1</Radio>
+        <Radio value="2">Option 2</Radio>
+      </RadioGroup>,
+    );
+
+    const group = getByRole("radiogroup");
+
+    expect(group).toHaveAttribute("aria-required", "true");
+
+    const radios = getAllByRole("radio");
+
+    expect(radios[0]).not.toHaveAttribute("required");
   });
 
   it("should work correctly with controlled value", () => {
@@ -204,11 +221,11 @@ describe("validation", () => {
   beforeAll(() => {
     user = userEvent.setup();
   });
-  describe("validationBehavior=native (default)", () => {
+  describe("validationBehavior=native", () => {
     it("supports isRequired", async () => {
       const {getAllByRole, getByRole, getByTestId} = render(
         <form data-testid="form">
-          <RadioGroup isRequired aria-label="favorite pet">
+          <RadioGroup isRequired aria-label="favorite pet" validationBehavior="native">
             <Radio value="dogs">Dogs</Radio>
             <Radio value="cats">Cats</Radio>
             <Radio value="dragons">Dragons</Radio>
@@ -244,6 +261,41 @@ describe("validation", () => {
       }
 
       expect(group).not.toHaveAttribute("aria-describedby");
+    });
+  });
+
+  describe("validationBehavior=aria", () => {
+    it("supports validate function", async () => {
+      const {getAllByRole, getByRole} = render(
+        <RadioGroup
+          aria-label="favorite pet"
+          defaultValue="dragons"
+          validate={(v) => (v === "dragons" ? "Too scary" : null)}
+          validationBehavior="aria"
+        >
+          <Radio value="dogs">Dogs</Radio>
+          <Radio value="cats">Cats</Radio>
+          <Radio value="dragons">Dragons</Radio>
+        </RadioGroup>,
+      );
+
+      const group = getByRole("radiogroup");
+
+      expect(group).toHaveAttribute("aria-describedby");
+      expect(group).toHaveAttribute("aria-invalid", "true");
+      expect(
+        document.getElementById(group.getAttribute("aria-describedby") as string),
+      ).toHaveTextContent("Too scary");
+
+      const radios = getAllByRole("radio") as HTMLInputElement[];
+
+      for (let input of radios) {
+        expect(input.validity.valid).toBe(true);
+      }
+
+      await user.click(radios[0]);
+      expect(group).not.toHaveAttribute("aria-describedby");
+      expect(group).not.toHaveAttribute("aria-invalid");
     });
   });
 });