+++ packages/components/skeleton/tsup.config.ts
@@ -4,4 +4,5 @@ export default defineConfig({
   clean: true,
   target: "es2019",
   format: ["cjs", "esm"],
+  banner: {js: '"use client";'},
 });