+++ packages/core/system/src/provider.tsx
@@ -7,6 +7,7 @@ import {RouterProvider} from "@react-aria/utils";
 import {OverlayProvider} from "@react-aria/overlays";
 import {useMemo} from "react";
 import {CalendarDate} from "@internationalized/date";
+import {MotionGlobalConfig} from "framer-motion";
 
 import {ProviderContext} from "./provider-context";
 
@@ -14,6 +15,14 @@ export interface NextUIProviderProps
   extends Omit<ModalProviderProps, "children">,
     ProviderContextProps {
   children: React.ReactNode;
+  /**
+   * Controls whether `framer-motion` animations are skipped within the application.
+   * This property is automatically enabled (`true`) when the `disableAnimation` prop is set to `true`,
+   * effectively skipping all `framer-motion` animations. To retain `framer-motion` animations while
+   * using the `disableAnimation` prop for other purposes, set this to `false`. However, note that
+   * animations in NextUI Components are still omitted if the `disableAnimation` prop is `true`.
+   */
+  skipFramerMotionAnimations?: boolean;
   /**
    * The locale to apply to the children.
    * @default "en-US"
@@ -38,6 +47,10 @@ export const NextUIProvider: React.FC<NextUIProviderProps> = ({
   children,
   navigate,
   useHref,
+  disableAnimation = false,
+  disableRipple = false,
+  skipFramerMotionAnimations = disableAnimation,
+  validationBehavior = "aria",
   locale = "en-US",
   defaultDates = {
     minDate: new CalendarDate(1900, 1, 1),
@@ -56,10 +69,26 @@ export const NextUIProvider: React.FC<NextUIProviderProps> = ({
     );
   }
 
-  const context = useMemo<ProviderContextProps>(
-    () => ({createCalendar, defaultDates}),
-    [createCalendar, defaultDates?.maxDate, defaultDates?.minDate],
-  );
+  const context = useMemo<ProviderContextProps>(() => {
+    if (disableAnimation && skipFramerMotionAnimations) {
+      MotionGlobalConfig.skipAnimations = true;
+    }
+
+    return {
+      createCalendar,
+      defaultDates,
+      disableAnimation,
+      disableRipple,
+      validationBehavior,
+    };
+  }, [
+    createCalendar,
+    defaultDates?.maxDate,
+    defaultDates?.minDate,
+    disableAnimation,
+    disableRipple,
+    validationBehavior,
+  ]);
 
   return (
     <ProviderContext value={context}>