+++ packages/core/theme/src/components/modal.ts
@@ -32,16 +32,6 @@ const modal = tv({
       "z-50",
       "overflow-x-auto",
       "justify-center",
-      //  mobile animation vars
-      "[--scale-enter:100%]",
-      "[--scale-exit:100%]",
-      "[--slide-enter:0px]",
-      "[--slide-exit:80px]",
-      // tablet/desktop animation vars
-      "sm:[--scale-enter:100%]",
-      "sm:[--scale-exit:103%]",
-      "sm:[--slide-enter:0px]",
-      "sm:[--slide-exit:0px]",
     ],
     base: [
       "flex",
@@ -173,6 +163,22 @@ const modal = tv({
       outside: {
         wrapper: "items-start sm:items-start overflow-y-auto",
         base: "my-16",
+      },
+    },
+    disableAnimation: {
+      false: {
+        wrapper: [
+          //  mobile animation vars
+          "[--scale-enter:100%]",
+          "[--scale-exit:100%]",
+          "[--slide-enter:0px]",
+          "[--slide-exit:80px]",
+          // tablet/desktop animation vars
+          "sm:[--scale-enter:100%]",
+          "sm:[--scale-exit:103%]",
+          "sm:[--slide-enter:0px]",
+          "sm:[--slide-exit:0px]",
+        ],
       },
     },
   },