+++ packages/components/kbd/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextui-org/kbd
 
+## 2.0.29
+
+### Patch Changes
+
+- Updated dependencies [[`e3afa4789`](https://github.com/nextui-org/nextui/commit/e3afa4789a1ac0fa929b2acaca5bd9c520567ab8), [`1109baea6`](https://github.com/nextui-org/nextui/commit/1109baea6ac6aa3feb2be90ef065f61b2c2a06a9)]:
+  - @nextui-org/system-rsc@2.1.2
+
 ## 2.0.28
 
 ### Patch Changes